+++ Projeto 2/common/io.h
@@ -1,5 +1,6 @@
 #ifndef COMMON_IO_H
 #define COMMON_IO_H
+#include <stddef.h>
 
 /// Parses an unsigned integer from the given file descriptor.
 /// @param fd The file descriptor to read from.
@@ -25,6 +26,6 @@ int print_str(int fd, const char *str);
 /// @param str The string to write.
 /// @param size The size of the string to write.
 /// @return 0 if the string was written successfully, 1 otherwise.
-int print_str_size(int fd, const char *str, int size);
+int print_str_size(int fd, const char *str, size_t size);
 
 #endif  // COMMON_IO_H
